requirements.txt
# TTS
pyttsx3

# STT
pyaudio
SpeechRecognition
vosk

# Audioplayer
python-vlc
keyboard

# UseCase3
spotipy
requests
geocoder

# FitBit
cherrypy
fitbit

#pytest
pytest
validators
